Add resource creation feature
The resources can now be created with some initial values.
In response, REST API can return a redirect to the newly
created resource. In that case resource creation routine will
load and return it.

Based on that, a few methods added to the `Port` object
to allow for adding VLANs.