Add `return_to` command processor feature
Prior to this change, returning from REPR `CommandProcessor.loop`
either terminates the CLI or returns to parent REPR loop.

This change allows loop caller to specify up to which command
processor should the call stack be unwound on exit.

The use-case is to support the CLIs which are not strictly
hierarchical, but allows the operator to jump between menus without
explicitly going through the parent menu.